move QueryOptimizeEngineTest.java
